requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.14.0
sqlparse==0.4.2
